Implemented three tests. Stuck on bug in database.
The third test, Successful POST request on /account, is not idempotent. The test user persists in the database after posting.

Either I need to figure out how to make the tests use a temporary database, or add a DELETE endpoint on /account/{id} to clean things up after the test is done.

Additonally, I am not sure if the cucumber tests are even run when running ./mvnw test. Only the contextLoads() method in RestApiApplicationTests is called.